fix(Établissement): différents correctifs et améliorations (#6955)
