+++ 2024-frontend/src/components/AppFieldDisplay.vue
@@ -23,7 +23,7 @@ const errorMessage = computed(() => hasError.value ? props.error?.join(' ') : ""
     <div class="fr-col-12 fr-col-md-7">
       <div class="ma-cantine--flex-start ma-cantine--flex-top fr-mb-1w">
         <AppRawHtml :html="valueFormatted" />
-        <DsfrTooltip v-if="tooltip" :content="tooltip" class="app-field-display__tooltip fr-ml-1w" on-hover="true" title="Infobulle">
+        <DsfrTooltip v-if="tooltip" :content="tooltip" class="app-field-display__tooltip fr-ml-1w" :on-hover="true" title="Infobulle">
           <VIcon name="ri-information-line" />
         </DsfrTooltip>
       </div>
