+++ 2024-frontend/src/components/CanteenButtonClaim.vue
@@ -2,7 +2,7 @@
 import { ref } from "vue"
 import { useRouter } from "vue-router"
 import { useRootStore } from "@/stores/root"
-import canteensService from "@/services/canteens"
+import managersService from "@/services/managers.js"
 import urlService from "@/services/urls"
 
 const props = defineProps(["id"])
@@ -13,7 +13,7 @@ const router = useRouter()
 /* Claim a canteen */
 const claimCanteen = () => {
   loading.value = true
-  canteensService
+  managersService
     .claimCanteen(props.id)
     .then((response) => {
       loading.value = false
