+++ 2024-frontend/src/services/managers.js
@@ -52,9 +52,37 @@ const removeManager = (canteenId, email) => {
     .catch((e) => e)
 }
 
+const claimCanteen = (id) => {
+  return fetch(`/api/v1/canteens/${id}/claim/`, {
+    method: "POST",
+    headers: {
+      "X-CSRFToken": window.CSRF_TOKEN || "",
+    },
+  })
+    .then(verifyResponse)
+    .then((response) => response)
+    .catch((e) => e)
+}
+
+const teamJoinRequest = (id, userInfos) => {
+  return fetch(`/api/v1/canteens/${id}/teamJoinRequest/`, {
+    method: "POST",
+    headers: {
+      "X-CSRFToken": window.CSRF_TOKEN || "",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(userInfos),
+  })
+    .then(verifyResponse)
+    .then((response) => response)
+    .catch((e) => e)
+}
+
 export default {
   fetchManagers,
   fetchManagerInvitations,
   addManager,
   removeManager,
+  claimCanteen,
+  teamJoinRequest,
 }
