--- 2024-frontend/src/views/GestionnaireCantineGerer.vue
@@ -1,251 +1,0 @@
-<script setup>
-import { computedAsync } from "@vueuse/core"
-import { useRoute } from "vue-router"
-import urlService from "@/services/urls.js"
-import canteenService from "@/services/canteens.js"
-import sectorService from "@/services/sectors.js"
-import cantines from "@/data/cantines.json"
-import AppLinkRouter from "@/components/AppLinkRouter.vue"
-
-/* Router */
-const route = useRoute()
-
-/* Canteen informations */
-const canteenInfos = computedAsync(
-  async () => {
-    const id = urlService.getCanteenId(route.params.canteenUrlComponent)
-    const canteen = await canteenService.fetchCanteen(id)
-    const isGroup = canteen.productionType === "groupe"
-    const editableInfos = getEditableInfos(canteen)
-    const notEditableInfos = !isGroup ? getNotEditableInfos(canteen) : null
-    const groupRelatedInfos = canteen.groupe ? getGroupRelatedInfos(canteen) : null
-    const editPageName = isGroup ? "GestionnaireCantineGroupeModifier" : "GestionnaireCantineRestaurantModifier"
-    return { editable: editableInfos, notEditable: notEditableInfos, groupRelated: groupRelatedInfos, editPage: editPageName, id: id }
-  },
-  { editable: [], notEditable: [], groupRelated: [], editPage: "", id: "" }
-)
-
-const getEditableInfos = (canteenInfos) => {
-  // Canteen
-  const hasSiren = canteenInfos.sirenUniteLegale
-  const hasSiret = canteenInfos.siret
-  const isGroup = canteenInfos.productionType === "groupe"
-  const isSite = canteenInfos.productionType === "site"
-  const isSatellite = canteenInfos.productionType === "site_cooked_elsewhere"
-  const hasLineMinistry = canteenInfos.lineMinistry
-  const hasRestaurant = isSite || isSatellite
-  // Infos
-  const infos = []
-  // Required field for all canteen
-  const fieldsName = ["name"]
-  // SIRET or SIREN
-  if (hasSiren) fieldsName.push("sirenUniteLegale")
-  if (hasSiret) fieldsName.push("siret")
-  // Types
-  fieldsName.push("productionType", "managementType")
-  if (hasRestaurant) fieldsName.push("economicModel")
-  // Meals
-  fieldsName.push("dailyMealCount", "yearlyMealCount")
-  // Sectors
-  if (hasRestaurant) fieldsName.push("sectorList")
-  if (hasLineMinistry) fieldsName.push("lineMinistry")
-  // Central producer SIRET
-  if (isSatellite || isGroup) fieldsName.push("centralProducerSiret")
-  // Geolocation
-  if(hasSiren && !isGroup) fieldsName.push("city", "postalCode")
-  // Get canteen fields value
-  fieldsName.forEach((name) => {
-    infos.push({ name: name, value: canteenInfos[name] })
-  })
-  return infos
-}
-
-const getNotEditableInfos = (canteenInfos) => {
-  const infos = []
-  const fieldsName = []
-  // Fields auto for SIRET canteen only
-  if (canteenInfos.siret) fieldsName.push("city", "postalCode")
-  // Fields auto for all canteen
-  fieldsName.push("cityInseeCode", "patLibList", "epciLib", "departmentLib", "regionLib")
-  fieldsName.forEach((name) => {
-    infos.push({ name: name, value: canteenInfos[name] })
-  })
-  return infos
-}
-
-const getGroupRelatedInfos = (canteenInfos) => {
-  const infos = []
-  const fieldsName = ["name", "sirenUniteLegale"]
-  fieldsName.forEach((name) => {
-    infos.push({ name: name, value: canteenInfos.groupe[name] })
-  })
-  return infos
-}
-
-
-/* Prettify */
-const getPrettyName = (name) => {
-  if (typeof cantines[name] === "string") return cantines[name]
-  if (typeof cantines[name] === "object") return cantines[`${name}Name`] // TODO : refactor json file to group name and options props
-}
-
-const getPrettyValue = (info) => {
-  const { name, value } = info
-  let prettyValue = null
-  switch (true) {
-    case name === "lineMinistry": {
-      prettyValue = getMinistrieName(value)
-      break
-    }
-    case name === "sectorList": {
-      prettyValue = getSectorsNames(value)
-      break
-    }
-    case typeof cantines[name] === "object": {
-      const index = cantines[name].findIndex((option) => option.value === value)
-      if (index >= 0) prettyValue = cantines[name][index].label
-      break
-    }
-    case value && value.length === 0: {
-      prettyValue = ""
-      break
-    }
-    case typeof value === "object" && value !== null: {
-      prettyValue = value.join(", ")
-      break
-    }
-    default: {
-      prettyValue = value
-    }
-  }
-  return prettyValue || "Non renseigné"
-}
-
-/* Sectors and ministry */
-const sectors = computedAsync(async () => {
-  return await sectorService.getSectors()
-}, [])
-const ministries = computedAsync(async () => {
-  return await sectorService.getMinistries()
-}, [])
-
-const getSectorsNames = (canteenSectorsIds) => {
-  const filteredSectors = sectors.value.filter((sector) => canteenSectorsIds.includes(sector.value))
-  const sectorsName = filteredSectors.map((filter) => filter.name)
-  return sectorsName.join(" ; ")
-}
-
-const getMinistrieName = (canteenMinistrySlug) => {
-  const index = ministries.value.findIndex((ministry) => ministry.value === canteenMinistrySlug)
-  return index >= 0 ? ministries.value[index].name : "Erreur sur le champ, contacter le support"
-}
-</script>
-
-<template>
-  <section>
-    <div class="fr-col-12 fr-col-md-8 fr-mb-3w">
-      <h1>{{ route.meta.title }}</h1>
-    </div>
-    <div class="fr-grid-row fr-grid-row--gutters">
-      <div class="fr-col-12" :class="{'fr-col-md-6': canteenInfos.notEditable}">
-        <div class="fr-card fr-p-2w fr-p-md-6w">
-          <h2 class="fr-h6">Informations renseignées</h2>
-          <p>Informations renseignées lors de la création de votre établissement.</p>
-          <ul class="ma-cantine--flex-grow">
-            <li v-for="info in canteenInfos.editable" :key="info.name">
-              <p class="fr-mb-0">
-                <span class="fr-text--bold">{{ getPrettyName(info.name) }} :</span>
-                {{ getPrettyValue(info) }}
-              </p>
-            </li>
-          </ul>
-          <div class="fr-grid-row fr-grid-row--center fr-mt-2w">
-            <router-link
-              class="ma-cantine--unstyled-link"
-              :to="{
-                name: canteenInfos.editPage,
-                params: { canteenUrlComponent: route.params.canteenUrlComponent },
-              }"
-            >
-              <DsfrButton secondary label="Modifier" />
-            </router-link>
-          </div>
-        </div>
-      </div>
-      <div v-if="canteenInfos.notEditable" class="fr-col-12 fr-col-md-6">
-        <div class="fr-card fr-p-2w fr-p-md-6w">
-          <h2 class="fr-h6">Informations générées</h2>
-          <p>
-            À partir des informations renseignées lors de la création de votre établissement, nous avons pu en générer
-            de nouvelles en croisant vos données avec d'autres référentiels :
-            <a href="https://france-pat.fr" target="_blank">France PAT</a>
-            et
-            <a href="https://annuaire-entreprises.data.gouv.fr" target="_blank">l'annuaire des entreprises</a>
-          </p>
-          <ul class="ma-cantine--flex-grow">
-            <li v-for="info in canteenInfos.notEditable" :key="info.name">
-              <p class="fr-mb-0">
-                <span class="fr-text--bold">{{ getPrettyName(info.name) }} :</span>
-                {{ getPrettyValue(info) }}
-              </p>
-            </li>
-          </ul>
-          <DsfrHighlight
-            class="fr-ml-0 fr-mb-0 fr-mt-2w ma-cantine--no-mb-p"
-            text="Ces informations ne sont pas modifiables, si vous remarquez une erreur merci de nous contacter."
-          />
-        </div>
-      </div>
-      <div v-if="canteenInfos.groupRelated" class="fr-col-12">
-        <div class="fr-card fr-p-2w fr-p-md-6w">
-          <h2 class="fr-h6">Groupe de restaurants satellites</h2>
-          <p>
-            Le gestionnaire du groupe de restaurants satellites a ajouté votre établissement.<br/>
-            Cela lui permet de réaliser une déclaration unique pour laquelle le montant total des achats du groupe est ensuite réparti automatiquement entre chaque restaurant satellite, au prorata de son nombre de couverts annuels.<br/>
-          </p>
-          <ul class="ma-cantine--flex-grow">
-            <li v-for="info in canteenInfos.groupRelated" :key="info.name">
-              <p class="fr-mb-0">
-                <span class="fr-text--bold">{{ getPrettyName(info.name) }} :</span>
-                {{ getPrettyValue(info) }}
-              </p>
-            </li>
-          </ul>
-          <DsfrHighlight
-            class="fr-ml-0 fr-mb-0 fr-mt-2w ma-cantine--no-mb-p"
-            text="Si vous remarquez une erreur ou souhaitez ne plus être associer au groupe, merci de nous contacter."
-          />
-        </div>
-      </div>
-      <div class="fr-col-12">
-        <div class="fr-card fr-p-2w fr-p-md-6w">
-          <h2 class="fr-h6">Identifiant (ID)</h2>
-          <p>Numéro identifiant attribué automatiquement lors de la création d'un établissement sur la plateforme <em>ma cantine</em>.
-            <br/>
-            Il est unique et non modifiable.
-          </p>
-          <ul>
-            <li>
-              <p class="fr-mb-0">
-                <span class="fr-text--bold">ID :</span>
-                {{ canteenInfos.id }}
-              </p>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </div>
-  </section>
-  <section class="fr-container fr-background-alt--red-marianne fr-p-3w fr-mt-3w fr-grid-row fr-grid-row--center">
-    <div class="fr-col-12 fr-col-lg-7 fr-background-default--grey fr-p-2w fr-p-md-7w">
-      <h2 class="fr-h5 fr-text-default--error">
-        <span class="mdi mdi-archive"></span>
-        Archiver cet établissement
-      </h2>
-      <p class="fr-mb-0">
-        Vous ne souhaitez plus faire apparaître cet établissement sur la plateforme <em>ma cantine</em>, vous pouvez l’archiver
-        <AppLinkRouter :to="{ name: 'GestionnaireCantineArchiver' }" title="en cliquant ici" />
-      </p>
-    </div>
-  </section>
-</template>
