+++ 2024-frontend/src/views/GestionnaireCantineTeledeclarations.vue
